webcontroller er blevet splittet i 3, for bedre overblik. Single responsibility principle.
